scide: switch CompletionMenu class to its own files
This patch switch the CompletionMenu class from the autocompleter.cpp file to
its own files in order to set it as a Q_OBJECT and use its own signals. This
patch also registers the completion_menu.hpp file into the ide_moc_hdr list.

The aim of this patch is to prepare the autocompleter help: the display of
additionnal help (description, summary, examples) when completing classes.
