+++ editors/sc-ide/CMakeLists.txt
@@ -61,6 +61,7 @@ set ( ide_moc_hdr
     widgets/code_editor/line_indicator.hpp
     widgets/code_editor/overlay.hpp
     widgets/code_editor/autocompleter.hpp
+    widgets/code_editor/completion_menu.hpp
     widgets/settings/dialog.hpp
     widgets/settings/general_page.hpp
     widgets/settings/sclang_page.hpp
@@ -112,6 +113,7 @@ set ( ide_src
     widgets/code_editor/line_indicator.cpp
     widgets/code_editor/overlay.cpp
     widgets/code_editor/autocompleter.cpp
+    widgets/code_editor/completion_menu.cpp
     widgets/settings/dialog.cpp
     widgets/settings/general_page.cpp
     widgets/settings/sclang_page.cpp
